requirements.txt
ultralytics>=8.2.0
opencv-python
torch>=2.0.0
torchvision>=0.15.0
numpy
matplotlib

django-allauth
PyJWT
python-decouple
ezgmail
cryptography
django-cors-headers
nanoid

fastapi==0.115.0
uvicorn[standard]==0.32.0
openai>=1.51.0
pydantic==2.9.2
pydantic-settings==2.6.0
jinja2==3.1.4
python-dotenv==1.0.1
langchain-tavily==0.2.0
langchain 
langchain-openai 
langchain-google-genai
loguru
google-generativeai

